requirements.txt
colorama==0.4.4
wheel==0.34.2
setuptools==52.0.0
pip==20.3.4
pyyaml==5.3.1
